Fix issue #3
The pixel mouse button callback can now distinguish between different
mouse buttons. The left mouse button acts the same as it did before, it
sets the cell that the mouse is hovering over. The right mouse button
now will unset the cell that the mouse is hovering over. And the middle
mouse button will toggle the cell that the mouse is over.